style updated in cplex parse log; corrected solve carnival - correct params list; generalising options checks - all solvers are checked in the same way. Not working code (yet).

+++ R/cleanup_carnival.R
@@ -10,7 +10,7 @@ cleanupCarnival <- function(carnivalOptions){
     resultFile <- carnivalOptions$filenames$resultFile
     keepLpFiles <- carnivalOptions$keepLPFiles
     
-    if(!keepLpFiles & file.exists(lpFile)){
+    if(!keepLpFiles && file.exists(lpFile)){
       file.remove(lpFile)
     }
